s18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
- import { Component, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Message } from '../models/message';
 import { Pagination } from '../models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,13 +17,13 @@ export class MessagesComponent implements OnInit {
   pageSize = 5
   loading = false
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages()
   }
 
-  loadMessages(){
+  loadMessages() {
     this.loading = true
     this.messageService.getMessages(this.pageNumber, this.pageSize, this.container).subscribe(respose => {
       this.messages = respose.result
@@ -31,14 +32,18 @@ export class MessagesComponent implements OnInit {
     })
   }
 
-  pageChanged(event: any){
+  pageChanged(event: any) {
     this.pageNumber = event.page
     this.loadMessages()
   }
 
-  deleteMessage(id: number){
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id ===id), 1)
+  deleteMessage(id: number) {
+    this.confirmService.confrim('Confirm dete message', 'This cannot be undone').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1)
+        })
+      }
     })
   }
 }
